// src/lib/shared/shared.type.ts
export enum ITEM_TYPE {
  SOFT_SHELL_CLAM = 'soft_shell_clam',
  ARCTIC_SURFCLAM = 'arctic_surfclam',
  GREENLAND_COCKLE = 'greenland_cockle',
  COMPACT_SNOW = 'compact_snow',
  ICE = 'ice',
  BLUE_ICE = 'blue_ice',
  SNOWMAN = 'snowman',
  ICE_HELMET = 'ice_helmet',
  TUSK_BLADES = 'tusk_blades'
}